refactor: mv wao-core/pkg -> wao-metrics-adapter/pkg
