+++ wao-loadbalancer/_src/pkg/proxy/nftables/wao.go
@@ -36,9 +36,9 @@ import (
 
 	// wao
 	waov1beta1 "github.com/waok8s/waok8s/wao-core/api/wao/v1beta1"
-	waoclient "github.com/waok8s/waok8s/wao-core/pkg/client"
-	waometrics "github.com/waok8s/waok8s/wao-core/pkg/metrics"
-	"github.com/waok8s/waok8s/wao-core/pkg/predictor"
+	waoclient "github.com/waok8s/waok8s/wao-metrics-adapter/pkg/client"
+	waometrics "github.com/waok8s/waok8s/wao-metrics-adapter/pkg/metrics"
+	"github.com/waok8s/waok8s/wao-metrics-adapter/pkg/predictor"
 )
 
 // Copied from wao-scheduler.
