+++ wao-metrics-adapter/Makefile
@@ -13,9 +13,9 @@ all: gen test build image
 gen:
 	go generate ./...
 
-.PHONY: test
-test: gen
-	go test -v -cover -race ./...
+# .PHONY: test
+# test: gen
+# 	go test -v -cover -race ./...
 
 .PHONY: image
 image: gen
@@ -28,3 +28,27 @@ push:
 .PHONY: build
 build: gen
 	go build $(GO_BUILD_ARGS) -o bin/adapter cmd/adapter/main.go
+
+#########################################
+# envtest copied from wao-core/Makefile #
+#########################################
+
+.PHONY: test
+test: envtest
+	KUBEBUILDER_ASSETS="$(shell $(ENVTEST) use $(ENVTEST_K8S_VERSION) --bin-dir $(LOCALBIN) -p path)" go test ./... -coverprofile cover.out
+
+# ENVTEST_K8S_VERSION refers to the version of kubebuilder assets to be downloaded by envtest binary.
+ENVTEST_K8S_VERSION = 1.30.0
+
+## Location to install dependencies to
+LOCALBIN ?= $(shell pwd)/bin
+$(LOCALBIN):
+	mkdir -p $(LOCALBIN)
+
+## Tool Binaries
+ENVTEST ?= $(LOCALBIN)/setup-envtest
+
+.PHONY: envtest
+envtest: $(ENVTEST) ## Download envtest-setup locally if necessary.
+$(ENVTEST): $(LOCALBIN)
+	test -s $(LOCALBIN)/setup-envtest || GOBIN=$(LOCALBIN) go install sigs.k8s.io/controller-runtime/tools/setup-envtest@latest
