+++ wao-metrics-adapter/pkg/predictor/interface.go
