+++ wao-metrics-adapter/pkg/predictor/redfish/api/api.go
