+++ wao-metrics-adapter/pkg/predictor/redfish/api/estimator-client.gen.go
