+++ wao-metrics-adapter/pkg/predictor/redfish/api/estimator-types.gen.go
