+++ wao-metrics-adapter/pkg/predictor/redfish/api/oapi-codegen-client.yaml
