+++ wao-metrics-adapter/pkg/predictor/redfish/api/openapi.yaml
